+++ src/Xamarin.Android.Build.Debugging.Tasks/Xamarin.Android.Common.Debugging.targets
@@ -367,21 +367,17 @@ Copyright (C) 2016 Xamarin. All rights reserved.
       AdbTarget="$(AdbTarget)"
       DiagnosticLogging="$(_FastDeploymentDiagnosticLogging)"
       FastDevFiles="@(_FilteredFastDevFiles);@(_ResolvedConfigFiles)"
-      FastDevToolPath="$(MSBuildThisFileDirectory)\lib"
       EmbedAssembliesIntoApk="$(EmbedAssembliesIntoApk)"
       PrimaryCpuAbi="$(_PrimaryCpuAbi)"
-      ToolsAbi="$(_FastDevToolsAbi)"
       PreserveUserData="$(AndroidPreserveUserData)"
       PackageName="$(_AndroidPackage)"
       PackageFile="$(_ApkToInstall)"
       ReInstall="$(_ReInstall)"
-      ToolVersion="$(AndroidFastDeploymentToolVersion)"
       AdbToolPath="$(AdbToolPath)"
       AdbToolExe="$(AdbToolExe)"
       AdbPushCompressionAlgorithm="$(AndroidFastDeploymentAdbCompressionAlgorithm)"
       AppFileTransferMode="$(_AndroidFastDeployAppFileTransferMode)"
       UploadFlagFile="$(_UploadFlag)"
-      UsingAndroidNETSdk="$(UsingAndroidNETSdk)"
       UserID="$(AndroidDeviceUserId)"
       IsTestOnly="$(_AndroidIsTestOnlyPackage)"
       IntermediateOutputPath="$(IntermediateOutputPath)"
